requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.1.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.1.5
django-rest-passwordreset==1.2.1
django-seed==0.3.1
django-smtp-ssl==1.0
djangorestframework==3.13.1
Faker==13.14.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
prompt-toolkit==3.0.30
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
toposort==1.7
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
